requirements.txt
streamlit>=1.24.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.86.0
python-dateutil>=2.8.2
pytz>=2023.3
google-cloud-core>=2.3.2
protobuf>=4.21.12
requests>=2.28.2